feat: 자기 글 리스트 조회 api

+++ src/main/java/com/example/board/board/dto/GetBoardListResponseDto.java
@@ -14,5 +14,5 @@ import java.util.List;
 @AllArgsConstructor
 public class GetBoardListResponseDto {
     private List<GetBoardDto> boardList;
-    private int totalCount;
+    private long totalCount;
 }
